# Refactor traverseValidationA' (perf).