--- a/src/FsToolkit.ErrorHandling/List.fs
+++ b/src/FsToolkit.ErrorHandling/List.fs
@@ -6,12 +6,12 @@ module List =
   let rec private traverseResultM' (state : Result<_,_>) (f : _ -> Result<_,_>) xs =
     match xs with
     | [] -> state
-    | x :: xs -> 
+    | x :: xs ->
       let r = result {
         let! y = f x
         let! ys = state
         return ys @ [y]
-      }  
+      }
       match r with
       | Ok _ -> traverseResultM' r f xs
       | Error _ -> r
@@ -19,22 +19,22 @@ module List =
   let rec private traverseAsyncResultM' (state : Async<Result<_,_>>) (f : _ -> Async<Result<_,_>>) xs =
     match xs with
     | [] -> state
-    | x :: xs -> 
+    | x :: xs ->
       async {
         let! r = asyncResult {
           let! ys = state
           let! y = f x
           return ys @ [y]
-        }  
+        }
         match r with
-        | Ok _ -> 
+        | Ok _ ->
           return! traverseAsyncResultM' (Async.singleton r) f xs
         | Error _ -> return r
       }
 
   let traverseResultM f xs =
     traverseResultM' (Ok []) f xs
-  
+
   let sequenceResultM xs =
     traverseResultM id xs
 
@@ -43,20 +43,20 @@ module List =
 
   let sequenceAsyncResultM xs =
     traverseAsyncResultM id xs
-  
+
 
   let rec private traverseResultA' state f xs =
     match xs with
     | [] -> state
     | x :: xs ->
-      let fR = 
+      let fR =
         f x |> Result.mapError List.singleton
       match state, fR with
-      | Ok ys, Ok y -> 
+      | Ok ys, Ok y ->
         traverseResultA' (Ok (ys @ [y])) f xs
-      | Error errs, Error e -> 
+      | Error errs, Error e ->
         traverseResultA' (Error (errs @ e)) f xs
-      | Ok _, Error e | Error e , Ok _  -> 
+      | Ok _, Error e | Error e , Ok _  ->
         traverseResultA' (Error e) f xs
 
   let rec private traverseAsyncResultA' state f xs =
@@ -67,11 +67,11 @@ module List =
         let! s = state
         let! fR = f x |> AsyncResult.mapError List.singleton
         match s, fR with
-        | Ok ys, Ok y -> 
+        | Ok ys, Ok y ->
           return! traverseAsyncResultA' (AsyncResult.retn (ys @ [y])) f xs
-        | Error errs, Error e -> 
+        | Error errs, Error e ->
           return! traverseAsyncResultA' (AsyncResult.returnError (errs @ e)) f xs
-        | Ok _, Error e | Error e , Ok _  -> 
+        | Ok _, Error e | Error e , Ok _  ->
           return! traverseAsyncResultA' (AsyncResult.returnError  e) f xs
       }
 
@@ -82,16 +82,19 @@ module List =
     traverseResultA id xs
 
   let rec traverseValidationA' state f xs =
-    match xs with
-    | [] -> state
-    | x  :: xs -> 
+    match state, xs with
+    | Ok items, [] ->
+      Ok (List.rev items)
+    | errors, [] ->
+      errors
+    | _, x :: xs ->
       let fR = f x
       match state, fR with
-      | Ok ys, Ok y -> 
-        traverseValidationA' (Ok (ys @ [y])) f xs
-      | Error errs1, Error errs2 -> 
+      | Ok ys, Ok y ->
+        traverseValidationA' (Ok (y :: ys)) f xs
+      | Error errs1, Error errs2 ->
         traverseValidationA' (Error (errs2 @ errs1)) f xs
-      | Ok _, Error errs | Error errs, Ok _  -> 
+      | Ok _, Error errs | Error errs, Ok _ ->
         traverseValidationA' (Error errs) f xs
 
   let traverseValidationA f xs =