@isaacabraham 2021-02-25
@TheAngryByrd Something I neglected to mention earlier - there might be other instances of this elsewhere in the codebase. Doing a search for `@ [` shows the following other occurrences:

```
src\FsToolkit.ErrorHandling\List.fs:
  12          let! ys = state
  13:         return ys @ [y]
  14        }

  26            let! y = f x
  27:           return ys @ [y]
  28          }

  55        | Ok ys, Ok y ->
  56:         traverseResultA' (Ok (ys @ [y])) f xs
  57        | Error errs, Error e ->

  70          | Ok ys, Ok y ->
  71:           return! traverseAsyncResultA' (AsyncResult.retn (ys @ [y])) f xs
  72          | Error errs, Error e ->

src\FsToolkit.ErrorHandling.JobResult\List.fs:
  13            let! y = f x
  14:           return ys @ [y]
  15          }  

  37          | Ok ys, Ok y -> 
  38:           return! traverseJobResultA' (JobResult.retn (ys @ [y])) f xs
  39          | Error errs, Error e -> 
```

These are probably good candidates for optimisation as they probably suffer from the same issue.